a1/CSC320/viscomp/requirements.txt
numpy==1.23.1
pandas
pillow
matplotlib
opencv-python
scipy
sourcedefender
